Sprite image url validity check performance improvement
