+++ packages/react-components/src/components/VideoGallery/utils/videoGalleryLayoutUtils.ts
@@ -4,6 +4,8 @@
 import { useCallback, useRef } from 'react';
 import { smartDominantSpeakerParticipants } from '../../../gallery';
 import { VideoGalleryParticipant, VideoGalleryRemoteParticipant } from '../../../types';
+/* @conditional-compile-remove(reaction) */
+import { ReactionResources } from '../../..';
 /* @conditional-compile-remove(gallery-layouts) */
 import { VideoGalleryLayout } from '../../VideoGallery';
 
@@ -263,4 +265,24 @@ export const useOrganizedParticipants = (args: OrganizedParticipantsArgs): Organ
   /* @conditional-compile-remove(pinned-participants) */
   return _useOrganizedParticipantsWithFocusedParticipants(args);
   return _useOrganizedParticipants(args);
+};
+
+/* @conditional-compile-remove(reaction) */
+/**
+ * @private
+ */
+export const getEmojiResource = (reactionName: string, reactionResources: ReactionResources): string | undefined => {
+  switch (reactionName) {
+    case 'like':
+      return reactionResources.likeReaction?.url;
+    case 'heart':
+      return reactionResources.heartReaction?.url;
+    case 'laugh':
+      return reactionResources.laughReaction?.url;
+    case 'applause':
+      return reactionResources.applauseReaction?.url;
+    case 'surprised':
+      return reactionResources.surprisedReaction?.url;
+  }
+  return '';
 };
